# packages/client/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserRole": {
				"type": "string",
				"enum": [
					"Admin",
					"Teacher",
					"Student"
				]
			},
			"UserRecord": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TagData": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"featured": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"featured"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectGraphRecord": {
				"properties": {
					"videoId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"objective": {
						"type": "string"
					},
					"assignments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/TagData"
						},
						"type": "array"
					},
					"levelStart": {
						"type": "number",
						"format": "double"
					},
					"levelEnd": {
						"type": "number",
						"format": "double"
					},
					"public": {
						"type": "boolean"
					},
					"collaborative": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"publishedAt": {
						"type": "string",
						"format": "date-time"
					},
					"shared": {
						"type": "boolean"
					},
					"shareName": {
						"type": "string"
					},
					"sharePassword": {
						"type": "string"
					},
					"shareExpiresAt": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserRecord"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/UserRecord"
						},
						"type": "array"
					}
				},
				"required": [
					"videoId",
					"title",
					"host",
					"objective",
					"assignments",
					"tags",
					"levelStart",
					"levelEnd",
					"public",
					"collaborative",
					"id",
					"userId",
					"publishedAt",
					"shared",
					"shareName",
					"sharePassword",
					"shareExpiresAt",
					"user",
					"members"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Extra": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"ProjectResult": {
				"properties": {
					"tags": {
						"items": {},
						"type": "array"
					},
					"id": {
						"type": "string"
					},
					"videoId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"assignments": {
						"items": {},
						"type": "array"
					},
					"publishedAt": {
						"type": "string",
						"format": "date-time"
					},
					"objective": {
						"type": "string"
					},
					"levelStart": {
						"type": "number",
						"format": "double"
					},
					"levelEnd": {
						"type": "number",
						"format": "double"
					},
					"public": {
						"type": "boolean"
					},
					"collaborative": {
						"type": "boolean"
					},
					"shared": {
						"type": "boolean"
					},
					"shareName": {
						"type": "string"
					},
					"shareExpiresAt": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"sharePassword": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"extra": {
						"$ref": "#/components/schemas/Extra"
					}
				},
				"required": [
					"tags",
					"id",
					"videoId",
					"userId",
					"title",
					"description",
					"host",
					"assignments",
					"publishedAt",
					"objective",
					"levelStart",
					"levelEnd",
					"public",
					"collaborative",
					"shared",
					"shareName",
					"shareExpiresAt",
					"sharePassword",
					"extra"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectCreateData": {
				"properties": {
					"videoId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"objective": {
						"type": "string"
					},
					"assignments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/TagData"
						},
						"type": "array"
					},
					"levelStart": {
						"type": "number",
						"format": "double"
					},
					"levelEnd": {
						"type": "number",
						"format": "double"
					},
					"public": {
						"type": "boolean"
					},
					"collaborative": {
						"type": "boolean"
					}
				},
				"required": [
					"videoId",
					"title",
					"host",
					"objective",
					"assignments",
					"tags",
					"levelStart",
					"levelEnd",
					"public",
					"collaborative"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@celluloid/server",
		"version": "2.0.0",
		"description": "Celluloid backend",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/projects": {
			"get": {
				"operationId": "GetProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProjectGraphRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResult"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectCreateData"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}